Start to implement renewal process and refactor parts of the signup process.
