+++ app/controllers/renewal/base_controller.rb
@@ -1,21 +1,13 @@
-module Signup
+module Renewal
   class BaseController < ApplicationController
+    before_action :authenticate_user!
+
     before_action :load_steps
     before_action :set_page_title
 
     layout "steps"
 
     private
-
-    def load_member
-      if session[:member_id] && session[:timeout] && session[:timeout] > Time.current - 15.minutes
-        @member = Member.find(session[:member_id])
-      else
-        reset_session
-        flash[:error] = "Your session expired. Please come into the library to complete signup."
-        redirect_to signup_url
-      end
-    end
 
     def load_steps
       agreement = Document.agreement
@@ -28,6 +20,14 @@ module Signup
       ]
     end
 
+    def load_member
+      @member = current_member
+    end
+
+    def set_page_title
+      @page_title = "Membership Renewal"
+    end
+
     def activate_step(step_id)
       @steps.each do |step|
         if step.id == step_id
@@ -36,10 +36,6 @@ module Signup
           break
         end
       end
-    end
-
-    def set_page_title
-      @page_title = "New Member Signup"
     end
   end
 end
