+++ app/controllers/renewal/confirmations_controller.rb
@@ -1,0 +1,8 @@
+module Renewal
+  class ConfirmationsController < BaseController
+    def show
+      activate_step(:complete)
+      @amount = Money.new(session[:amount]) if session.key?(:amount)
+    end
+  end
+end
