+++ app/controllers/renewal/documents_controller.rb
@@ -1,0 +1,14 @@
+module Renewal
+  class DocumentsController < BaseController
+    def rules
+      @document = Document.borrow_policy
+      activate_step(:rules)
+    end
+
+    def agreement
+      @document = Document.agreement
+      @acceptance = AgreementAcceptance.new
+      activate_step(:agreement)
+    end
+  end
+end
