+++ app/controllers/renewal/home_controller.rb
@@ -1,0 +1,7 @@
+module Renewal
+  class HomeController < BaseController
+    def index
+      @hide_steps = true
+    end
+  end
+end
