+++ app/controllers/renewal/redemptions_controller.rb
@@ -1,0 +1,35 @@
+module Renewal
+  class RedemptionsController < BaseController
+    before_action :load_member
+
+    def new
+      activate_step(:payment)
+      @form = GiftMembershipRedemptionForm.new
+    end
+
+    def create
+      @form = GiftMembershipRedemptionForm.new(form_params)
+      if @form.valid?
+        Membership.transaction do
+          @membership = Membership.create_for_member(@member)
+          @gift_membership = GiftMembership.find(@redemption.gift_membership_id)
+          @gift_membership.redeem(@membership)
+        end
+
+        MemberMailer.with(member: @member).welcome_message.deliver_later
+
+        reset_session
+        redirect_to signup_confirmation_url
+      else
+        activate_step(:payment)
+        render :new
+      end
+    end
+
+    private
+
+    def form_params
+      params.require(:gift_membership_redemption_form).permit(:code)
+    end
+  end
+end
