--- app/lib/signup/payment.rb
@@ -1,14 +1,0 @@
-module Signup
-  class Payment
-    include ActiveModel::Model
-
-    attr_accessor :amount_dollars
-
-    validates_numericality_of :amount_dollars,
-      greater_than_or_equal_to: 1, less_than_or_equal_to: 500
-
-    def amount
-      Money.new(amount_dollars.to_i * 100)
-    end
-  end
-end
