--- app/views/layouts/signup.html.erb
@@ -1,39 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Circulate</title>
-    <%= csrf_meta_tags %>
-    <%= csp_meta_tag %>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, shrink-to-fit=no">
-    <meta http-equiv="x-ua-compatible" content="ie=edge">
-    <% unless request.get? && !form_request? # prevent turbolinks for caching form redisplays, which will often contain validation messages %>
-      <meta name="turbolinks-cache-control" content="no-preview">
-    <% end %>
-    <%= javascript_pack_tag 'application', 'data-turbolinks-track': 'reload' %>
-    <%= stylesheet_pack_tag 'styles', media: 'all', 'data-turbolinks-track': 'reload' %>
-    <% if Rails.env.production? %>
-      <script async defer data-domain="app.chicagotoollibrary.org" src="https://plausible.io/js/plausible.js"></script>
-    <% end %>
-  </head>
-
-  <body>
-    <div class="container grid-lg signup app">
-      <%= flash_message :success  %>
-      <%= flash_message :warning  %>
-      <%= flash_message :error  %>
-
-      <div class="columns">
-        <div class="column col-12">
-          <div class="signup-logo <%= "on-step" if @active_step %>">
-            <%= image_pack_tag "logo.png" %> 
-            <h1 class="title">New Member Signup</h1>
-          </div>
-          <%= render 'signup/steps', steps: @steps unless @hide_steps %>
-        </div>
-      </div>
-
-      <%= yield %>
-    </div>
-
-  </body>
-</html>
