+++ app/views/renewal/confirmations/_instructions.html.erb
@@ -1,0 +1,22 @@
+
+<% if @amount %>
+  <p><strong>Your payment of <%= @amount.format %> has been processed and applied to your account.</strong></p>
+<% end %>
+<p>To activate your membership, please bring the following with you to the library the first time you visit:</p>
+<ul>
+  <li><strong>Proof of your address</strong><br>This needs to include your name and current address in Chicago. Utility bills, leases, and other official mail work well for this.</li>
+  <li><strong>A photo ID</strong><br>This needs to show your name and a photo and can be from the state, your job, school, etc.</li>
+</ul>
+<p>We're located in Bridgeport at:<br>
+  1048 W. 37th Street Suite 102<br>
+  Chicago, IL 60609</p>
+
+<div class="info">
+  <p>You can <a href="https://app.chicagotoollibrary.org/users/sign_in">login to your account</a> at any time and update your contact information, browse our tool inventory, and request your first tool pick-up!</p>
+</div>
+
+<div class="warning">
+  <p><strong>
+    We are open <em>by appointment only</em> in response to COVID-19. <a href="https://chicagotoollibrary.org/covid-19">Please check our COVID-19 page for more information.</a>
+  </strong></p>
+</div>
