+++ test/controllers/signup/payments_controller_test.rb
@@ -27,7 +27,7 @@ module Signup
       }]
 
       SquareCheckout.stub :new, mock_checkout do
-        post signup_payments_url, params: {signup_payment: {amount_dollars: "12"}}
+        post signup_payments_url, params: {membership_payment_form: {amount_dollars: "12"}}
       end
 
       assert_redirected_to "https://squareup.com/checkout/12345"
@@ -45,7 +45,7 @@ module Signup
       mock_checkout.expect :checkout_url, mock_result, [Hash]
 
       SquareCheckout.stub :new, mock_checkout do
-        post signup_payments_url, params: {signup_payment: {amount_dollars: "12"}}
+        post signup_payments_url, params: {membership_payment_form: {amount_dollars: "12"}}
       end
 
       assert_redirected_to "http://www.example.com/signup/payments/new"
